requirements.txt
ollama
faiss-cpu==1.8.0
tqdm



